% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AClass_230815_2.R
\name{batch.nano.eval.test}
\alias{batch.nano.eval.test}
\title{Batch evaluation of classification results}
\usage{
batch.nano.eval.test(
  prefix,
  use_class = NULL,
  Prob_range = seq(from = 0, to = 1, by = 0.01),
  prob = "Avg_Probability",
  training_memberships_path,
  GeoMean_thres = NULL,
  run_dir_path
)
}
\arguments{
\item{prefix}{Prefix string to identify result files.}

\item{use_class}{custom class output order. Required field.}

\item{Prob_range}{vector of probability intervals to be used in analysis. Default 0 to 1 by 0.01.}

\item{prob}{column name for probability present in *_test_summary.txt file. Default "Avg_Probability".}

\item{training_memberships_path}{no header, expects nano_filename in column 1 and and Class in column 2.}

\item{GeoMean_thres}{Housekeeping gene geometric mean threshold to be considered in analysis. Default NULL for no filtering and is same as using 0.}

\item{run_dir_path}{expects multiple test results folders stored within this path.}
}
\value{
No object returned. Evaluation results are written to Excel and RDS files in each subfolder and as a combined summary.
}
\description{
Batch processing \code{nano.eval.test()} and create an overall summary for all runs within dir
}
