% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AClass_230815_2.R
\name{batch.process.raw}
\alias{batch.process.raw}
\title{Batch Processing Function of NanoString Data}
\usage{
batch.process.raw(
  work_path = getwd(),
  raw_dir_path,
  keep_file_path = NULL,
  omit_file_path = NULL,
  prefix = NULL,
  SampleContent = "housekeeping.geo.mean",
  mode = "combined"
)
}
\arguments{
\item{work_path}{Directory where the output folder will be created. Defaults to current working directory.}

\item{raw_dir_path}{expects multiple raw directories stored within this path.}

\item{keep_file_path}{Optional. File listing samples to keep (one sample ID per row, no header).}

\item{omit_file_path}{Optional. File listing samples to omit (one sample ID per row, no header).}

\item{prefix}{Optional string to prefix output directories and files.}

\item{SampleContent}{Normalization method passed to \code{nano.norm()}. Default is \code{"housekeeping.geo.mean"}.}

\item{mode}{Processing mode that controls how data are loaded. \code{"batch"} option reads each directory in raw_dir_path and normalize separately before merging. \code{"combined"} option reads all data within raw_dir_path recursively and processes as one. Default is \code{"combined"}.}

\item{raw_path}{Path to NanoString data file or folder (default *NormalizedData.csv) file. Required}

\item{recursive_read}{Logical. Whether to recursively scan directories for input files. Default is \code{FALSE}.}
}
\value{
A normalized and pre-processed AClass object containing merged results across multiple batches or combined dataset.
}
\description{
A batch wrapper for \code{process.raw()} that allows multiple NanoString datasets
to be processed either individually (batch mode) or together (combined mode)
All parameters matche with \code{process.raw()} except raw_dir_path and mode
}
