% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AClass_230815_2.R
\name{classify.data}
\alias{classify.data}
\title{Perform ensemble classification using pre-trained model}
\usage{
classify.data(
  work_path = NULL,
  data,
  prefix,
  training_model_obj,
  alg_list = c("rf", "glmnet", "pam", "nb", "knn"),
  keep_file_path = NULL,
  omit_file_path = NULL,
  out_path = NULL,
  thres_geomean = 100,
  report_type = "Summary"
)
}
\arguments{
\item{work_path}{Directory where the output folder will be created. Default is NULL}

\item{data}{Pre-processed AClass object}

\item{prefix}{Optional string to prefix output directories and files.}

\item{training_model_obj}{Pre-trained model}

\item{alg_list}{Vector of algorithms in the training_model_obj to be used for classification. Default is \code{c("rf","glmnet","pam", "nb", "knn")}}

\item{keep_file_path}{Optional. File listing samples to keep (one sample ID per row, no header).}

\item{omit_file_path}{Optional. File listing samples to omit (one sample ID per row, no header).}

\item{out_path}{Alternative output location instead of run_id under work_path}

\item{thres_geomean}{threshold for report to be considered passing. Default is \code{100}. \code{NULL} indicate skipping threshold check.}

\item{report_type}{report format options. \code{Summary} or \code{Detailed} description. Refer to \code{nano.plot()}}
}
\value{
A classified AClass object with test results, probability scores, assigned groups, and create visualization outputs.
}
\description{
Wrapper function to classify pre-normalized expression data (NanoString or compatible) using a pre-trained AClass model. Wraps \code{nano.test}, \code{get.nano.test.results}, \code{nano.set.colour}, and \code{nano.plot}.
}
\examples{
data.obj <- classify.data(data = data.obj,prefix = "demo",training_model_obj = training_models)
}
