% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AClass_230815_2.R
\name{initialize.prj}
\alias{initialize.prj}
\title{Installs and loads required packages for AClass project. This includes both CRAN and Bioconductor packages,
and ensures all dependencies are available. Note: archived version of NanoStringNorm is utilized}
\usage{
initialize.prj(install_missing = TRUE)
}
\arguments{
\item{install_missing}{Logical. Whether to install missing packages automatically. Default is TRUE.}
}
\value{
Prints messages about package loading status.
}
\description{
Installs and loads required packages for AClass project. This includes both CRAN and Bioconductor packages,
and ensures all dependencies are available. Note: archived version of NanoStringNorm is utilized
}
