% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AClass_230815_2.R
\name{nano.eval.test}
\alias{nano.eval.test}
\title{Evaluate test results across probability thresholds}
\usage{
nano.eval.test(
  prefix,
  use_class = NULL,
  Prob_range = seq(from = 0, to = 1, by = 0.01),
  prob = "Avg_Probability",
  anno_table = NULL,
  training_memberships_path = NULL,
  GeoMean_thres = NULL,
  out_path = NULL,
  in_path = getwd(),
  recursive_read = FALSE
)
}
\arguments{
\item{prefix}{Prefix string to identify result files.}

\item{use_class}{Character vector specifying class order for factor alignment.}

\item{Prob_range}{Numeric vector of thresholds to evaluate (e.g., seq(0,1,0.01)).}

\item{prob}{Column name in *_test_summary.txt indicating prediction probability.}

\item{training_memberships_path}{Path to a tab-delimited file with two columns: sample ID (column 1) and ground truth class (column2).}

\item{GeoMean_thres}{Optional numeric threshold to filter on GeoMean values.}

\item{out_path}{Directory to write results. Defaults to working directory.}

\item{in_path}{Directory to read *_test_summary.txt files. Defaults to working directory.}

\item{recursive_read}{Logical, whether to search in_path recursively.}
}
\value{
A list with two entries: overall accuracy and confusion matrix.
}
\description{
This function aggregates and evaluates model predictions from *_test_summary.txt files fount in \code{in_path}, generating performance metrics across specified probability thresholds and producing confusion matrices (if \code{Group} column exist in dataframe) and accuracy summaries. Note: \code{confusionMatrix()} from the \code{caret} package requires more than two classes to compute multiclass metrics reliably.
}
\examples{
# Run evaluation (requires *_test_summary.txt and mapping file)
# nano.eval.test(
#   prefix = "demo",
#   use_class = c("SHH", "TYR", "MYC"),
#   training_memberships_path = "sample_to_class.txt"
# )
}
