% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AClass_230815_2.R
\name{nano.load}
\alias{nano.load}
\title{Looks for Nanostring Data with "NormalizedData.csv" and load as dataframe}
\usage{
nano.load(
  raw_path = getwd(),
  keep_file_path = "",
  omit_file_path = "",
  recursive_read = FALSE
)
}
\arguments{
\item{raw_path}{path of the directory where raw nanostring data is kept. Read recursively.}

\item{omit_file_path}{path and tab-delimited list of sample names to be omitted. One per row.}

\item{recursive_read}{read everything in folder. This may have large impact depending on variations between runs. Recommend FALSE by default.}
}
\value{
dataframe of read data and number of samples loaded.
}
\description{
Looks for Nanostring Data with "NormalizedData.csv" and load as dataframe
}
