% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AClass_230815_2.R
\name{nano.MDS}
\alias{nano.MDS}
\title{General-purpose data MDS plotting.}
\usage{
nano.MDS(
  prefix,
  data,
  plot_type = c("boxplot", "plot", "ggplot", "ggplot_label", "ggplot_label_batch",
    "plotly", "plotly_batch"),
  data_name = c("norm.t", "train.data.main", "train.data.validate")
)
}
\arguments{
\item{prefix}{Character string used in plot titles.}

\item{data}{AClass object with normalized expression data.}

\item{plot_type}{Type of plot to generate. Options: "boxplot", "plot", "ggplot", "ggplot_label", "ggplot_label_batch", "plotly", "plotly_batch"}

\item{data_name}{Name of the data slot to use. Options: "norm.t", "train.data.main", or "train.data.validate".}
}
\value{
No object is returned; a plot (ggplot2 or plotly) is printed to the active device based on the selected type.
}
\description{
Requires nano.set.colour() to be run first to determine colour_code
}
\examples{
 library(limma) #plotMDS
 library(plotly)
 library(reshape2)
nano.MDS(prefix = project_prefix, data=train, plot_type = "ggplot",data_name = "norm.t")
}
