% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AClass_230815_2.R
\name{nano.plot}
\alias{nano.plot}
\title{Visualize classification results and QC metrics}
\usage{
nano.plot(
  prefix,
  data,
  prob = "Avg_Probability",
  thres_avg_prob = 0,
  thres_geomean = NULL,
  report_type = c("Summary", "Detailed"),
  print_report = FALSE,
  out_path = NULL
)
}
\arguments{
\item{prefix}{file prefix (optional)}

\item{prob}{"Avg_Probability" or "Avg_Cal_Probability"}

\item{thres_avg_prob}{Probability threshold below which predictions are considered low confidence. Related to "prob". Default to 0.}

\item{thres_geomean}{geomean threshold cut off below which will be considered fail. Default to NULL which indicates this test is skipped.}

\item{report_type}{report format options. "Summary" or "Detailed"}

\item{print_report}{binary option to print txt summary "test_summary_aggregate.txt" and "test_summary.txt"}

\item{out_path}{output path. When not provided out_path will be extracted from run_info (default)}
}
\description{
Generates PDF plots summarizing prediction confidence, model agreement, and sample quality control. Includes both detailed and summary report options with support for calibrated or raw probabilities.
}
\examples{
test.obj <- nano.plot(prefix = "demo", data = test.obj, report_type = "Summary", print_report = TRUE)
}
