% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AClass_230815_2.R
\name{nano.prenorm.qc}
\alias{nano.prenorm.qc}
\title{Prenorm Nanostring Data and produce report. Required to calculate HK_geomean}
\usage{
nano.prenorm.qc(data, code_class = "Housekeeping", prefix, out_path = NULL)
}
\arguments{
\item{code_class}{Default "Housekeeping".}

\item{prefix}{project prefix}

\item{raw_data}{Raw dataframe from nano.load.}
}
\value{
Report of the raw data and GeoMean table.
}
\description{
Prenorm Nanostring Data and produce report. Required to calculate HK_geomean
}
