% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AClass_230815_2.R
\name{nano.prep}
\alias{nano.prep}
\title{Transform and check dataframe for missing values}
\usage{
nano.prep(data)
}
\arguments{
\item{train.norm}{Normalized dataframe.}

\item{test.norm}{Normalized dataframe.}
}
\value{
??.
}
\description{
Transform and check dataframe for missing values
}
