% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AClass_230815_2.R
\name{nano.train}
\alias{nano.train}
\title{Prenorm Nanostring Data and produce report}
\usage{
nano.train(
  prefix,
  data,
  work_path,
  alg_list = c("rf", "glmnet", "pam", "nb", "knn"),
  probes_rank_path = NULL,
  min_test_features = 20,
  max_test_features = 30,
  c.method = "repeatedcv",
  c.repeats = 5,
  c.number = 10,
  out_path = NULL
)
}
\arguments{
\item{work_path}{project work path}

\item{alg_list}{list of alg}

\item{probes_rank_path}{path to probe ranking list. Given list sorted by p value. Looks for probes_list.txt in the working directory if probes_rank_path is missing.}

\item{min_test_features}{Minimum number of features tested.}

\item{max_test_features}{Maximum number of features tested.}

\item{out_path}{output path. When not provided out_path will be extracted from run_info (default) and work_path}

\item{train.norm}{Normalized dataframe.}
}
\value{
Optimal_Training_Attributes Full_Training_Attributes.txt performance_pdf
}
\description{
Prenorm Nanostring Data and produce report
}
