% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AClass_230815_2.R
\name{process.raw}
\alias{process.raw}
\title{Function to load, normalize and pre-process NanoString data}
\usage{
process.raw(
  work_path = getwd(),
  raw_path = NULL,
  keep_file_path = NULL,
  omit_file_path = NULL,
  prefix = NULL,
  SampleContent = "housekeeping.geo.mean",
  recursive_read = FALSE
)
}
\arguments{
\item{work_path}{Directory where the output folder will be created. Defaults to current working directory.}

\item{raw_path}{Path to NanoString data file or folder (default *NormalizedData.csv) file. Required}

\item{keep_file_path}{Optional. File listing samples to keep (one sample ID per row, no header).}

\item{omit_file_path}{Optional. File listing samples to omit (one sample ID per row, no header).}

\item{prefix}{Optional string to prefix output directories and files.}

\item{SampleContent}{Normalization method passed to \code{nano.norm()}. Default is \code{"housekeeping.geo.mean"}.}

\item{recursive_read}{Logical. Whether to recursively scan directories for input files. Default is \code{FALSE}.}
}
\value{
A normalized and pre-processed AClass object.
}
\description{
Wrapper function to process NanoString data by running \code{nano.load()}, \code{nano.prenorm.qc()}, \code{nano.prep()} and \code{nano.norm()}
}
