% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AClass_230815_2.R
\name{df2nano}
\alias{df2nano}
\title{Convert annotated data frame to AClass object format}
\usage{
df2nano(
  df,
  colour_code = NULL,
  add_to = c("train.data.main", "train.data.validate")
)
}
\arguments{
\item{df}{A data frame with a \code{Group} column and genes/features as columns and samples as rows.}

\item{colour_code}{Colour code data frame. Must contain \code{Group} and \code{Group_Colour} columns.}

\item{add_to}{Either a character ("train.data.main" or "train.data.validate") or a data frame with two columns specifying which samples go to the training or validation set. Default is \code{train.data.main}}
}
\value{
A list representing a formatted transcriptomic object compatible with AClass tools.
}
\description{
Converts a data frame into an AClass-style list object with support for
training/validation splits and group color annotations. This function is useful for importing custom transcriptomic data into the AClass workflow. By default, the input data frame is assigned to \code{train.data.main}.
}
