% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AClass_230815_2.R
\name{nano.load}
\alias{nano.load}
\title{Load NanoString NormalizedData.csv files and return structured raw object}
\usage{
nano.load(
  raw_path = getwd(),
  keep_file_path = "",
  omit_file_path = "",
  recursive_read = FALSE
)
}
\arguments{
\item{raw_path}{path of the directory where raw nanostring data is kept. Read recursively.}

\item{keep_file_path}{Optional. File listing samples to keep (one sample ID per row, no header).}

\item{omit_file_path}{Optional. File listing samples to omit (one sample ID per row, no header).}

\item{recursive_read}{Logical. Whether to recursively search subdirectories. May introduce run-to-run variation. Default is FALSE.}
}
\value{
AClass object, which is a list of dataframe with read data and number of samples loaded.
}
\description{
Loads all files with "NormalizedData" in the filename and ".csv" extension, starting at row 16.
}
\examples{
raw.obj <- nano.load(
  raw_path = "data/",
  keep_file_path = "samples_keep.txt",
  omit_file_path = "samples_omit.txt"
)
}
