% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AClass_230815_2.R
\name{nano.test}
\alias{nano.test}
\title{Prenorm Nanostring Data and produce report}
\usage{
nano.test(
  prefix,
  training_model_obj,
  data,
  alg_list = NULL,
  min_test_features = NULL,
  max_test_features = NULL,
  out_path = NULL
)
}
\arguments{
\item{prefix}{output prefix e.g. paste(project,format(Sys.time(), "\%Y-\%m-\%d_\%H\%M"), sep = "_")}

\item{data}{test data.}

\item{alg_list}{list of alg}

\item{min_test_features}{Minimum number of features tested.}

\item{max_test_features}{Maximum number of features tested.}

\item{out_path}{output path. When not provided out_path will be extracted from run_info (default)}

\item{training_model_list}{Trained_models.}
}
\value{
testing_results_summary_groups_score
}
\description{
Prenorm Nanostring Data and produce report
}
