% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AClass_230815_2.R
\name{remap.subgroups}
\alias{remap.subgroups}
\title{Remap subgroup labels in a slot of AClass object}
\usage{
remap.subgroups(
  data,
  slot = "test_results_full",
  column = "Class",
  map = "Torchia2016_to_Ho2019"
)
}
\arguments{
\item{data}{AClass object that has been analyzed with \code{nano.test()}.}

\item{slot}{Character. Slot name within AClass object to be modified. Default is "test_results_full".}

\item{column}{Character. Column within the slot to apply remapping. Default is "Class".}

\item{map}{Character. Mapping scheme to use. Currently supports only "Torchia2016_to_Ho2019".}
}
\value{
Modified AClass object with updated group labels.
}
\description{
Internal helper function to remap subgroup labels from Torchia et al. 2016 to Ho et al. 2019 consensus names.
Intended to be called after \code{nano.test()}, when \code{test_results_full} is present.
}
\keyword{internal}
